Updated and finalised examples-sim-CRM, minor adjustments to related scripts
Updated examples-sim-CRM to include a new plotting function plot_CRM in utilities.py, and to simulate and save parameters as pickle and curves to csv with 2 species and 2 resources. Consistent with layout and functions used in gLV examples.
Moved relevant functions to examples-bayes-CRM and explore-CRM, and reflormatting of sim_gLV,